// apps/api/src/generated/prisma/enums.ts
export const OrganizationMemberRole = {
  OWNER: 'OWNER',
  ADMIN: 'ADMIN',
  RECRUITER: 'RECRUITER',
  STAFF: 'STAFF'
} as const

export type OrganizationMemberRole = (typeof OrganizationMemberRole)[keyof typeof OrganizationMemberRole]
// ...
export const OrganizationVerificationDecisionAction = {
  NEEDS_INFORMATION: 'NEEDS_INFORMATION',
  VERIFIED: 'VERIFIED',
  REJECTED: 'REJECTED'
} as const

export type OrganizationVerificationDecisionAction = (typeof OrganizationVerificationDecisionAction)[keyof typeof OrganizationVerificationDecisionAction]